Fixed warning from covariance matrix
The method used to fit the phi-value (scipy.optimize.curve_fit) automatically attempts to calculate a covariance matrix from the fit values.  As we are fitting only a single value, AND two of the fits (where the parents originate) come back with an exact fit, this was throwing an error message about being unable to find the covariance.  That message is now suppressed via the warnings package.